Play a square wave! (make your ears bleed)
Day 008

Interestingly, if the frequency divides evenly into the sample rate,
it sounds smooth, but otherwise it sounds awful and scratchy. This
may be because it's still buggy like Casey mentioned. Everything looks
fine to me stepping through the code, but then again I don't yet have
a good grasp of how it's really supposed to work.